Split macro tests into multiple files
The MacroSystemTests file had grown into a giant test-all file. Split the tests into multiple files, grouped by the role of the macro they are testing.
